+++ app/layout.tsx
@@ -1,4 +1,6 @@
-
+import "@/styles/globals.css";
+import { Inter } from "@next/font/google";
+const inter = Inter({ subsets: ["latin"] });
 
 export default function RootLayout({
   children,
